pyproject.toml
[project]
name = "dc-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "discord-py[voice]>=2.6.4",
    "spotipy>=2.25.1",
    "yt-dlp>=2025.10.22",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
